feat: Eliminate Code Duplication and Unify Border Functionality
